fix: move end2end tests to their own test package
